Move basic Parameters code into its own small framework. Make RemoteParameters depend on it.
