+++ Parameters/Parameters.h
@@ -1,0 +1,19 @@
+//
+//  Parameters.h
+//  Parameters
+//
+//  Created by Avi Cieplinski on 6/3/20.
+//  Copyright © 2020 Avi Cieplinski. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for Parameters.
+FOUNDATION_EXPORT double ParametersVersionNumber;
+
+//! Project version string for Parameters.
+FOUNDATION_EXPORT const unsigned char ParametersVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <Parameters/PublicHeader.h>
+
+
