+++ RemoteParameters/RemoteParameters.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import Parameters
 import MultipeerConnectivity
 
 struct RemoteParameterAttribute {
